backend/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.20
PyMySQL==1.0.3
python-dotenv==1.0.0
gunicorn==20.1.0
